+++ core/interface/orchestrator_action/src/lib.rs
@@ -1,0 +1,11 @@
+mod error;
+mod registry;
+mod traits;
+
+pub use self::error::ActionAlreadyRegistered;
+pub use self::registry::OrchestratorActionRegistry;
+pub use self::registry::OrchestratorActionRegistryBuilder;
+pub use self::traits::OrchestratorAction;
+
+#[cfg(feature = "test_api")]
+pub use self::registry::TestRegistryClearGuard;